Fixed UI and filter issues across multiple reports, ensured consistent date formatting, and verified stability during data parsing

+++ lib/StaffModule/screen/Reports/ReportScreens/Expiring_Insurance.dart
@@ -272,7 +272,7 @@ class _ExpiringInsuranceState extends State<ExpiringInsurance> {
                     lease.expirationDate != null
                         ? dateProvider.formatCurrentDate(lease.expirationDate!)
                         : '',
-                    formatCurrency(lease.liabilityCoverage),
+                    formatCurrency(lease.liabilityCoverage.toDouble()),
                     lease.tenantDetails != null
                         ? lease.tenantDetails
                             ?.map((tenant) =>
@@ -365,7 +365,7 @@ class _ExpiringInsuranceState extends State<ExpiringInsurance> {
           : "");
       sheet
           .getRangeByIndex(2 + i, 6)
-          .setText(formatCurrency(lease.liabilityCoverage));
+          .setText(formatCurrency(lease.liabilityCoverage.toDouble()));
       String? tenantNames = lease.tenantDetails != null
           ? lease.tenantDetails
               ?.map((tenant) =>
@@ -457,7 +457,7 @@ class _ExpiringInsuranceState extends State<ExpiringInsurance> {
         lease.expirationDate != null
             ? dateProvider.formatCurrentDate(lease.expirationDate!)
             : '',
-        formatCurrency(lease.liabilityCoverage),
+        formatCurrency(lease.liabilityCoverage.toDouble()),
         lease.tenantDetails != null
             ? lease.tenantDetails
                 ?.map((tenant) =>
